+++ src/util.h
@@ -11,6 +11,6 @@ float smooth(float t);
 float smoother(float t);
 float normalize(float value, float min, float max);
 float hash_posf(int x, int y);
-size_t hash_pos(int x, int y);
+uint32_t hash_pos(int x, int y);
 
 #endif
